app-backend: added utility to find assets to store
