+++ plugins/app-backend/package.json
@@ -35,6 +35,7 @@
     "@backstage/config": "^0.1.12",
     "@backstage/types": "^0.1.1",
     "@types/express": "^4.17.6",
+    "globby": "^11.0.0",
     "express": "^4.17.1",
     "express-promise-router": "^4.1.0",
     "fs-extra": "9.1.0",
